Move MeteredData to own OrchestrationStarter
