+++ source/IncomingMessages.Infrastructure/ProcessManager/IRequestProcessOrchestrationStarter.cs
@@ -25,8 +25,4 @@ public interface IRequestProcessOrchestrationStarter
     Task StartRequestAggregatedMeasureDataOrchestrationAsync(
         InitializeAggregatedMeasureDataProcessDto initializeProcessDto,
         CancellationToken cancellationToken);
-
-    Task StartForwardMeteredDataForMeasurementPointOrchestrationAsync(
-        InitializeMeteredDataForMeasurementPointMessageProcessDto initializeProcessDto,
-        CancellationToken cancellationToken);
 }
